+++ NEWS.md
@@ -3,6 +3,9 @@
 ## New functions
 * `tg_channels_search()`- Search telegram channel.
 
+## Bug fixes
+* Fix `tg_get_token()`, previously the function only looked for the token in the options, ignoring the environment variable (#1)
+
 ## Other
 * `tg_get_token()` return error when API token doesn't set.